refactor : Lotto 인터페이스 분리 해제
